jobs/requirements.txt
pandas
numpy
google-cloud-bigquery
db-dtypes
nba_api
